Added checked functionality and confirm to clear all stored tasks

+++ app.js
@@ -7,27 +7,42 @@ const tableBody = document.getElementById('table-body');
 const clearBtn = document.getElementById('clear');
 
 // store the tasks in an array
-let taskList = ['Delete Me...'];
+let taskList = [
+  {
+    task: 'Delete me....',
+    checkClass: ''
+  },
+  {
+    task: 'Checked example',
+    checkClass: 'checked'
+  }
+];
 
 // load event listeners
-init();
-
-function init() {
+(function() {
   // document.addEventListener('DOMContentLoaded', getLocal());
   form.addEventListener('submit', addTask);
   clearBtn.addEventListener('click', clearTasks);
-  tableBody.addEventListener('click', deleteTask);
+  tableBody.addEventListener('click', deleteOrCheck);
   getLocal(taskList);
   taskList.forEach(writeTask);
+})();
+
+//new task constructor
+function Task(task, checkClass = '') {
+  this.task = task;
+  this.checkClass = checkClass;
 }
+
 
 //get task and push to taskList array
 function addTask(e) {
   if (taskInput.value === '') {
     alert('Please enter a task'); //change this for a better alert
   } else {
-    taskList.push(taskInput.value);
-    writeTask(taskInput.value);
+    taskList.push(new Task(taskInput.value));
+    //write the task with the last one added to array
+    writeTask(taskList[taskList.length - 1]);
     taskInput.value = '';
   }
   e.preventDefault();
@@ -37,38 +52,60 @@ function addTask(e) {
 function writeTask(task) {
   const tr = document.createElement('tr');
   const a = '<a class="delete is-pulled-right"></a>';
-  tr.innerHTML = `<td>${task}${a}</td>`;
+  tr.innerHTML = `<td class="${task.checkClass}">${task.task}${a}</td>`;
   tableBody.appendChild(tr);
   setLocal(taskList);
 }
 
-//delete a task
-function deleteTask(e) {
+//delete or check a task
+function deleteOrCheck(e) {
+  //delete task
   if (e.target.classList.contains('delete')) {
+    let taskText = e.target.parentElement.textContent;
     //remove it from the array
-    let taskToRemove = taskList.indexOf(e.target.parentElement.textContent);
-    taskList.splice(taskToRemove, 1);
-    //remove it from the table
+    for (i = 0; i < taskList.length; i++) {
+      if (taskList[i].task === taskText) {
+        taskList.splice(i, 1);
+      }
+    }
+    //remove from dom
     e.target.parentElement.parentElement.remove();
+  }
+  // check task
+  else if (e.target.localName === 'td') {
+    e.target.classList.toggle('checked');
+    let taskText = e.target.textContent;
+    for (i = 0; i < taskList.length; i++) {
+      if (taskList[i].task === taskText) {
+        if (taskList[i].checkClass == '') {
+          taskList[i].checkClass = 'checked';
+        } else {
+          taskList[i].checkClass = '';
+        }
+      }
+    }
   }
   setLocal(taskList);
 }
 
 //clear all tasks
 function clearTasks() {
-  //empty array
-  taskList = [];
-  //empty table
-  tableBody.innerHTML = '';
-  //clear local
-  localStorage.clear();
+  let confirmed =  confirm('Remove all stored tasks?');
+  if (confirmed === true){
+    //empty array
+    taskList = [];
+    //empty table
+    tableBody.innerHTML = '';
+    //clear local
+    localStorage.clear();
+  }
 }
 
-//add local storage
 //set local storage
 function setLocal(array) {
   localStorage.setItem('tasks', JSON.stringify(array));
 }
+
 //get local storage
 function getLocal(array) {
   if (localStorage.getItem('tasks') === null) {
